mac_desktop_app/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pandas==2.2.2
email-validator==2.2.0
pywebview==4.4.1
